perf: 添加BasePermissionsViewSet，支持搜索过滤
